+++ src/forward_model/modules/module.h
@@ -41,7 +41,7 @@ class Module{
     virtual void modifySpectrumGPU(
       const std::vector<double>& parameter,
       Atmosphere* atmosphere,
-      double* spectrum_gpu) = 0;
+      float* spectrum_gpu) = 0;
     size_t nbParameters() {return nb_parameters;}
   protected:
     size_t nb_parameters {};
