+++ src/forward_model/secondary_eclipse_bb/secondary_eclipse_bb.h
@@ -94,8 +94,8 @@ class OccultationBlackBodyModel : public ForwardModel{
       std::vector<std::vector<double>>& spectrum_obs);
     virtual bool calcModelGPU(
       const std::vector<double>& parameters,
-      double* spectrum,
-      std::vector<double*>& spectrum_obs);
+      float* spectrum,
+      std::vector<float*>& spectrum_obs);
     
     virtual void postProcess(
       const std::vector< std::vector<double> >& model_parameter,
@@ -131,16 +131,16 @@ class OccultationBlackBodyModel : public ForwardModel{
       const std::vector<double>& parameters);
 
     void calcOccultationGPU(
-      double* secondary_eclipse,
-      double* planet_spectrum,
-      const double* stellar_spectrum,
+      float* secondary_eclipse,
+      float* planet_spectrum,
+      const float* stellar_spectrum,
       const int nb_points,
       const double radius_ratio,
-      const double* albedo_contribution);
+      const float* albedo_contribution);
     
     void calcPlanetSpectrumGPU(
       const double planet_temperature,
-      double* spectrum_dev);
+      float* spectrum_dev);
 
     void postProcess(
       const OccultationBlackBodyPostConfig& post_process_config,
