+++ src/forward_model/stellar_spectrum/stellar_spectrum_grid.h
@@ -97,7 +97,7 @@ class StellarSpectrumGrid : public StellarSpectrumModel{
     
     virtual void calcFluxGPU(
       const std::vector<double>& parameter,
-      double* spectrum_gpu);
+      float* spectrum_gpu);
   protected:
     SpectralGrid* spectral_grid;
     std::vector<double> grid_wavelengths;
